Updated template to support Glue 3.0
