--- .github/workflows/scala.yml
@@ -1,19 +1,0 @@
-name: Scala CI
-
-on:
-  push:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up JDK 1.8
-      uses: actions/setup-java@v1
-      with:
-        java-version: 1.8
-    - name: Compile
-      run: sbt clean compile
